refactor: restructure Prisma integration with global PrismaModule
- Moved PrismaService to PrismaModule and defined it as global.
- Updated modules (auth, user) to use PrismaModule instead of directly injecting PrismaService.
