+++ src/modules/user/user.module.ts
@@ -1,11 +1,10 @@
 import { Module } from "@nestjs/common";
-import { PrismaService } from "src/prisma.service";
 import { UserService } from "./user.service";
 import { UserController } from "./user.controller";
 
 @Module({
     controllers: [UserController],
-    providers: [PrismaService, UserService],
+    providers: [UserService],
     exports: [UserService]
 })
